* commit 'e4be11fbe02cb555c2334a56d4933949bef552b0':
  Dont add unnecessary data to rollback query
  Fix priority complete handler for last priority
  Small filler fixes
  Fix materialized view order
  Fix api bugs
  Fix filler bugs
  Add indexes and foreign keys for new tables
  Update example config and remove all non contract apis
  Fix linting and building errors
  Update AtomicMarket and AtomicTools filler
  Implement new filler structure
  Update data deserialization + processing architecture
  Add data processing class

+++ definitions/tables/delphioracle_tables.sql
@@ -5,7 +5,7 @@ CREATE TABLE delphioracle_pairs (
     base_precision integer NOT NULL,
     quote_symbol character varying(12) NOT NULL,
     quote_precision integer NOT NULL,
-    median integer,
+    median bigint,
     median_precision integer NOT NULL,
     updated_at_time bigint NOT NULL,
     updated_at_block bigint NOT NULL,
